Added basic combat
The units now battle eachother in an intence and action packed matter!
No but really they now can attack and have an attack cooldown to maybe
wait for an animation or similar.
Also forgot to add a NavMesh last commit, but this time I didn't, so the
units can roam around on the battlefield!

- Added "Ally.prefab" prefab
- Changed "Enemy.prefab" prefab
- Changed "Hammer.prefab" prefab
- Changed "Projectile.prefab" prefab
- Changed "HammerController.cs" script
- Changed "Minion.cs" script
- Changed "Projectile.cs" script
- Changed "WackingArea.cs" script
- Added "NavMesh.asset"
- Changed physics layer collision settings