Now contains a main menu! :smile: :+1:
Previously there were no main menu, but fear no more 'cause I have
overengineered it and tried to make a cutscene-like menu.

I focused on making this with animations instead of programming it
because it gives you more possibilities. It may be a bit tideous but it
gives a new level of control. My goal for this is to remove the
animation events and make that automatic, at least. Don't know how I
would do that tho...

Also added in a fading so it fades to and from black when changing
scenes.